Resolved cyclic data access issue created by loading code from `background.js` inside `BrowserStorageManager` (which itself was loaded from `background.js`)
